requirements.txt
Django
Markdown
Pillow
PyJWT
anyascii
asgiref
beautifulsoup4
certifi
charset-normalizer
crispy-bootstrap5
crispy-tailwind
defusedxml
distlib
dj-database-url
django-auto-logout
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-modelcluster
django-permissionedforms
django-taggit
django-treebeard
djangorestframework
djangorestframework-simplejwt
draftjs-exporter
env
et-xmlfile
filelock
filetype
html5lib
idna
l18n
laces
openpyxl
pi
pillow-heif
pip
pipenv
platformdirs
psycopg2
psycopg2-binary
python-dotenv
pytz
requests
setuptools
six
soupsieve
sqlparse
telepath
typing-extensions
tzdata
urllib3
virtualenv
wagtail
webencodings
willow
